Add container stats to stack page
* refactor container stats into separate component
* add said component to the stack page as an additional tab
